==> code/mouse/notebooks/09_merge_public_datasets.ipynb <==
import os
import scanpy as sc
import pandas as pd
import numpy as np
import scipy.sparse as sp
import anndata as ad
import seaborn as sns

dir_path = '/mnt/storage/Daniele/atlases/mouse_public_raw/'

studies = os.listdir(dir_path)

## Carstens 2021

carstens_adata = sc.read_h5ad(f'{dir_path}/Carstens_2021/adata_merged_raw.h5ad')

carstens_adata.obs['Sample_unique'] = (carstens_adata.obs.Sample.astype(str) + "_" + carstens_adata.obs.Dataset.astype(str)).astype('category')


## Chen 2021

chen_2021_adata = sc.read_h5ad(f'{dir_path}/Chen_2021/adata_merged_raw.h5ad')

chen_2021_adata.obs['Sample_unique'] = (chen_2021_adata.obs.Sample.astype(str) + "_" + chen_2021_adata.obs.Dataset.astype(str)).astype('category')


## Chen 2024

chen_2024_adata = sc.read_h5ad(f'{dir_path}/Chen_2024/adata_merged_raw.h5ad')

chen_2024_adata.obs['Dataset'] = 'Chen_2024'

chen_2024_adata[:10,:10].X.todense()

chen_2024_adata.obs['Sample_unique'] = (chen_2024_adata.obs.Sample.astype(str) + "_" + chen_2024_adata.obs.Dataset.astype(str)).astype('category')


## Cornell 2021 miss raw counts

cornell_2021_adata = sc.read_h5ad(f'{dir_path}/Cornell_2021/adata_merged_raw.h5ad')

cornell_2021_adata[:10,:10].X.todense()

## Donahue 2024

donahue_adata = sc.read_h5ad(f'{dir_path}/Donahue_2024/adata_merged_raw.h5ad')

donahue_adata.obs_names[0][19:]

donahue_adata.obs['Dataset'] = 'Donahue_2024'

donahue_adata.obs['Sample_unique'] = (donahue_adata.obs.Sample.astype(str) + "_" + donahue_adata.obs.Dataset.astype(str)).astype('category')


donahue_adata[:10,:10].X.todense()

## Elyada 2020

elyada_adata = sc.read_h5ad(f'{dir_path}/Elyada_2020/adata_merged_raw.h5ad')

elyada_adata.obs['Sample_unique'] = (elyada_adata.obs.Sample.astype(str) + "_" + elyada_adata.obs.Dataset.astype(str)).astype('category')


elyada_adata[:10,:10].X.todense()

## Erdem 2024

erden_2024_adata = sc.read_h5ad(f'{dir_path}/Erden_2024/adata_merged_raw.h5ad')

erden_2024_adata.obs['Dataset'] = 'Erdem_2024'

erden_2024_adata.obs['Sample_unique'] = (erden_2024_adata.obs.Sample.astype(str) + "_" + erden_2024_adata.obs.Dataset.astype(str)).astype('category')


erden_2024_adata[:10,:10].X.todense()

## Veghin 2024

veghin_2024_adata = sc.read_h5ad(f'{dir_path}/Veghin_2024/adata_merged_raw.h5ad')

veghin_2024_adata.obs['Sample_unique'] = (veghin_2024_adata.obs.Sample.astype(str) + "_" + veghin_2024_adata.obs.Dataset.astype(str)).astype('category')


veghin_2024_adata[:10,:10].X.todense()

## Singhai 2024 miss raw counts

singhai_2024_adata = sc.read_h5ad(f'{dir_path}/Singhal_2024/adata_merged_raw.h5ad')

singhai_2024_adata.obs['Sample_unique'] = (singhai_2024_adata.obs.Sample.astype(str) + "_" + singhai_2024_adata.obs.Dataset.astype(str)).astype('category')


## Rupert 2025

rupert_2025_adata = sc.read_h5ad(f'{dir_path}/Rupert_2025/adata_merged_raw.h5ad')

rupert_2025_adata.obs['Sample_unique'] = (rupert_2025_adata.obs.Sample.astype(str) + "_" + rupert_2025_adata.obs.Dataset.astype(str)).astype('category')


rupert_2025_adata.X.A

## Hosein 2022

hosein_2022_adata = sc.read_h5ad(f'{dir_path}/Hosein_2022/adata_merged_raw.h5ad')

hosein_2022_adata.obs['Sample_unique'] = (hosein_2022_adata.obs.Sample.astype(str) + "_" + hosein_2022_adata.obs.Dataset.astype(str)).astype('category')


## Hosein 2019

hosein_2019_adata = sc.read_h5ad(f'{dir_path}/Hosein_2019/adata_merged_raw.h5ad')

hosein_2019_adata.obs['Sample_unique'] = (hosein_2019_adata.obs.Sample.astype(str) + "_" + hosein_2019_adata.obs.Dataset.astype(str)).astype('category')


## Han 2023

han_2023_adata = sc.read_h5ad('/mnt/storage/Daniele/atlases/mouse_public_raw/Han_2023/adata_merged_raw.h5ad')

han_2023_adata.obs['Sample_unique'] = (han_2023_adata.obs.Sample.astype(str) + "_" + han_2023_adata.obs.Dataset.astype(str)).astype('category')


han_2023_adata

han_2023_adata.X.A

# combine

adata = ad.concat([carstens_adata, chen_2021_adata, chen_2024_adata,donahue_adata,elyada_adata, erden_2024_adata,rupert_2025_adata,hosein_2022_adata, hosein_2022_adata,han_2023_adata], join = 'outer')

adata

adata.write_h5ad('/mnt/storage/Daniele/atlases/mouse/08_mouse_public_raw.h5ad')
